requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
phonenumbers==8.13.26
